+++ src/resolvers/high-order-resolvers/authenticate.ts
@@ -2,7 +2,7 @@ import { skip } from 'graphql-resolvers'
 
 import { IContext } from 'utils/apollo-server'
 
-export function isAuthenticated(root, args, { authUser }: IContext, info) {
-  if (!authUser) return new Error('Not authenticated.')
+export function isAuthenticated(root, args, { authUser, ERROR_TYPES }: IContext, info) {
+  if (!authUser) return new Error(ERROR_TYPES.UNAUTHENTICATED)
   else return skip
 }